Crate brotlic_sys

Structs

BrotliDecoderState

Opaque structure that holds decoder state.

BrotliEncoderPreparedDictionary

Opaque type for pointer to different possible internal structures containing dictionary prepared for the encoder

BrotliEncoderState

Opaque structure that holds encoder state.

BrotliSharedDictionary

Opaque structure that holds shared dictionary data.

Constants

BROTLI_DEFAULT_QUALITY

BROTLI_DEFAULT_WINDOW

BROTLI_FALSE

BROTLI_LARGE_MAX_WINDOW_BITS

BROTLI_MAX_INPUT_BLOCK_BITS

BROTLI_MAX_QUALITY

BROTLI_MAX_WINDOW_BITS

BROTLI_MIN_INPUT_BLOCK_BITS

BROTLI_MIN_QUALITY

BROTLI_MIN_WINDOW_BITS

BROTLI_TRUE

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_ALLOC_BLOCK_TYPE_TREES

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_ALLOC_CONTEXT_MAP

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_ALLOC_CONTEXT_MODES

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_ALLOC_RING_BUFFER_1

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_ALLOC_RING_BUFFER_2

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_ALLOC_TREE_GROUPS

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_COMPOUND_DICTIONARY

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_DICTIONARY_NOT_SET

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_BLOCK_LENGTH_1

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_BLOCK_LENGTH_2

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_CL_SPACE

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_CONTEXT_MAP_REPEAT

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_DICTIONARY

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_DISTANCE

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_EXUBERANT_META_NIBBLE

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_EXUBERANT_NIBBLE

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_HUFFMAN_SPACE

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_PADDING_1

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_PADDING_2

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_RESERVED

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_SIMPLE_HUFFMAN_ALPHABET

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_SIMPLE_HUFFMAN_SAME

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_TRANSFORM

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_FORMAT_WINDOW_BITS

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_INVALID_ARGUMENTS

BrotliDecoderErrorCode_BROTLI_DECODER_ERROR_UNREACHABLE

BrotliDecoderErrorCode_BROTLI_DECODER_NEEDS_MORE_INPUT

BrotliDecoderErrorCode_BROTLI_DECODER_NEEDS_MORE_OUTPUT

BrotliDecoderErrorCode_BROTLI_DECODER_NO_ERROR

BrotliDecoderErrorCode_BROTLI_DECODER_SUCCESS

BrotliDecoderParameter_BROTLI_DECODER_PARAM_DISABLE_RING_BUFFER_REALLOCATION

Disable “canny” ring buffer allocation strategy.

BrotliDecoderParameter_BROTLI_DECODER_PARAM_LARGE_WINDOW

Flag that determines if “Large Window Brotli” is used.

BrotliDecoderResult_BROTLI_DECODER_RESULT_ERROR

Decoding error, e.g. corrupted input or memory allocation problem.

BrotliDecoderResult_BROTLI_DECODER_RESULT_NEEDS_MORE_INPUT

Partially done; should be called again with more input.

BrotliDecoderResult_BROTLI_DECODER_RESULT_NEEDS_MORE_OUTPUT

Partially done; should be called again with more output.

BrotliDecoderResult_BROTLI_DECODER_RESULT_SUCCESS

Decoding successfully completed.

BrotliEncoderMode_BROTLI_MODE_FONT

Compression mode used in WOFF 2.0.

BrotliEncoderMode_BROTLI_MODE_GENERIC

Default compression mode.

BrotliEncoderMode_BROTLI_MODE_TEXT

Compression mode for UTF-8 formatted text input.

BrotliEncoderOperation_BROTLI_OPERATION_EMIT_METADATA

Emit metadata block to stream.

BrotliEncoderOperation_BROTLI_OPERATION_FINISH

Finalize the stream.

BrotliEncoderOperation_BROTLI_OPERATION_FLUSH

Produce output for all processed input.

BrotliEncoderOperation_BROTLI_OPERATION_PROCESS

Process input.

BrotliEncoderParameter_BROTLI_PARAM_DISABLE_LITERAL_CONTEXT_MODELING

Flag that affects usage of “literal context modeling” format feature.

BrotliEncoderParameter_BROTLI_PARAM_LARGE_WINDOW

Flag that determines if “Large Window Brotli” is used.

BrotliEncoderParameter_BROTLI_PARAM_LGBLOCK

Recommended input block size.

BrotliEncoderParameter_BROTLI_PARAM_LGWIN

Recommended sliding LZ77 window size.

BrotliEncoderParameter_BROTLI_PARAM_MODE

Tune encoder for specific input.

BrotliEncoderParameter_BROTLI_PARAM_NDIRECT

Recommended number of direct distance codes (NDIRECT).

BrotliEncoderParameter_BROTLI_PARAM_NPOSTFIX

Recommended number of postfix bits (NPOSTFIX).

BrotliEncoderParameter_BROTLI_PARAM_QUALITY

The main compression speed-density lever.

BrotliEncoderParameter_BROTLI_PARAM_SIZE_HINT

Estimated total input size for all ::BrotliEncoderCompressStream calls.

BrotliEncoderParameter_BROTLI_PARAM_STREAM_OFFSET

Number of bytes of input stream already processed by a different instance.

BrotliSharedDictionaryType_BROTLI_SHARED_DICTIONARY_RAW

Raw LZ77 prefix dictionary.

BrotliSharedDictionaryType_BROTLI_SHARED_DICTIONARY_SERIALIZED

Serialized shared dictionary.

Functions

BrotliDecoderAttachDictionary⚠

Adds LZ77 prefix dictionary, adds or replaces built-in static dictionary and transforms.

BrotliDecoderCreateInstance⚠

Creates an instance of ::BrotliDecoderState and initializes it.

BrotliDecoderDecompress⚠

Performs one-shot memory-to-memory decompression.

BrotliDecoderDecompressStream⚠

Decompresses the input stream to the output stream.

BrotliDecoderDestroyInstance⚠

Deinitializes and frees ::BrotliDecoderState instance.

BrotliDecoderErrorString⚠

Converts error code to a c-string.

BrotliDecoderGetErrorCode⚠

Acquires a detailed error code.

BrotliDecoderHasMoreOutput⚠

Checks if decoder has more output.

BrotliDecoderIsFinished⚠

Checks if decoder instance reached the final state.

BrotliDecoderIsUsed⚠

Checks if instance has already consumed input.

BrotliDecoderSetParameter⚠

Sets the specified parameter to the given decoder instance.

BrotliDecoderTakeOutput⚠

Acquires pointer to internal output buffer.

BrotliDecoderVersion⚠

Gets a decoder library version.

BrotliEncoderAttachPreparedDictionary⚠

Attaches a prepared dictionary of any type to the encoder. Can be used multiple times to attach multiple dictionaries. The dictionary type was determined by BrotliEncoderPrepareDictionary. Multiple raw prefix dictionaries and/or max 1 serialized dictionary with custom words can be attached.

BrotliEncoderCompress⚠

Performs one-shot memory-to-memory compression.

BrotliEncoderCompressStream⚠

Compresses input stream to output stream.

BrotliEncoderCreateInstance⚠

Creates an instance of ::BrotliEncoderState and initializes it.

BrotliEncoderDestroyInstance⚠

Deinitializes and frees ::BrotliEncoderState instance.

BrotliEncoderDestroyPreparedDictionary⚠

BrotliEncoderEstimatePeakMemoryUsage⚠

BrotliEncoderGetPreparedDictionarySize⚠

BrotliEncoderHasMoreOutput⚠

Checks if encoder has more output.

BrotliEncoderIsFinished⚠

Checks if encoder instance reached the final state.

BrotliEncoderMaxCompressedSize⚠

Calculates the output size bound for the given @p input_size.

BrotliEncoderPrepareDictionary⚠

Prepares a shared dictionary from the given file format for the encoder.

BrotliEncoderSetParameter⚠

Sets the specified parameter to the given encoder instance.

BrotliEncoderTakeOutput⚠

Acquires pointer to internal output buffer.

BrotliEncoderVersion⚠

Gets an encoder library version.

BrotliSharedDictionaryAttach⚠

Attaches dictionary to a given instance of ::BrotliSharedDictionary.

BrotliSharedDictionaryCreateInstance⚠

Creates an instance of ::BrotliSharedDictionary.

BrotliSharedDictionaryDestroyInstance⚠

Deinitializes and frees ::BrotliSharedDictionary instance.

Type Aliases

BROTLI_BOOL

BrotliDecoderErrorCode

Error code for detailed logging / production debugging.

BrotliDecoderParameter

Options to be used with ::BrotliDecoderSetParameter.

BrotliDecoderResult

Result type for ::BrotliDecoderDecompress and ::BrotliDecoderDecompressStream functions.

BrotliEncoderMode

Options for ::BROTLI_PARAM_MODE parameter.

BrotliEncoderOperation

Operations that can be performed by streaming encoder.

BrotliEncoderParameter

Options to be used with ::BrotliEncoderSetParameter.

BrotliSharedDictionaryType

Input data type for ::BrotliSharedDictionaryAttach.

brotli_alloc_func

Allocating function pointer type.

brotli_free_func

Deallocating function pointer type.