Function brotlic_sys::BrotliEncoderPrepareDictionary

pub unsafe extern "C" fn BrotliEncoderPrepareDictionary(
    type_: BrotliSharedDictionaryType,
    data_size: usize,
    data: *const u8,
    quality: c_int,
    alloc_func: brotli_alloc_func,
    free_func: brotli_free_func,
    opaque: *mut c_void
) -> *mut BrotliEncoderPreparedDictionary

Prepares a shared dictionary from the given file format for the encoder.

@p alloc_func and @p free_func @b MUST be both zero or both non-zero. In the case they are both zero, default memory allocators are used. @p opaque is passed to @p alloc_func and @p free_func when they are called. @p free_func has to return without doing anything when asked to free a NULL pointer.

@param type type of dictionary stored in data @param data_size size of @p data buffer @param data pointer to the dictionary data @param quality the maximum Brotli quality to prepare the dictionary for, use BROTLI_MAX_QUALITY by default @param alloc_func custom memory allocation function @param free_func custom memory free function @param opaque custom memory manager handle